server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
# Pin to a currently available psycopg binary build compatible with Python 3.13
psycopg[binary]==3.2.12
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
# Pin bcrypt to a version compatible with passlib to avoid backend errors
bcrypt==3.2.2
